Bugfix/not deleteable alarm

+++ src/main/java/at/schulgong/controller/RingtoneController.java
@@ -211,12 +211,17 @@ public class RingtoneController {
      * @return deleted ringtone
      */
     @DeleteMapping("/{id}")
-    ResponseEntity<RingtoneDTO> deleteRingtone(@PathVariable long id) {
+    ResponseEntity<?> deleteRingtone(@PathVariable long id) {
         if (ringtoneRepository.existsById(id)) {
             RingtoneDTO ringtoneDTO = one(id);
-            ringtoneRepository.deleteById(id);
-            deleteAudioFile(ringtoneDTO.getPath());
-            return ResponseEntity.noContent().build();
+            if(!ringtoneDTO.getName().toLowerCase().equals("alarm")){
+              ringtoneRepository.deleteById(id);
+              deleteAudioFile(ringtoneDTO.getPath());
+              return ResponseEntity.noContent().build();
+            }else {
+              return ResponseEntity.status(400)
+                .body("Der Alarm kann nicht gelöscht werden!");
+            }
         } else {
             throw new EntityNotFoundException(id, Config.RINGTONE.getException());
         }
